+++ changelog.md
@@ -5,7 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
-- Add all `Operating System` trove classifiers.
+- Add all `Operating System ::` trove classifiers.
+- Add all platform names as package keywords.
 
 ## [4.0.0 (2025-10-21)](https://github.com/kdeldycke/extra-platforms/compare/v3.2.3...v4.0.0)